Add files via upload
Please find the attached jar file and code as an attached,

How to execute.

1) Make sure java virtual machine should installed in your local machine
2) copy the jar file in the folder.
3) execute the command from the same folder. ( java -jar BmoProject.jar )

O notation is good for binary search. Algorithm performance depend on the data amount input , output and processing time.

If the algorithm lies in a good zone that would be best otherwise exponential one is bad zone . O notation log is better after some-point it steady. Please find the screen shot for time of processing.